Added x86_64 smp

+++ kernel/src/common/drivers/apic.c
@@ -1,3 +1,5 @@
+#include <stdint.h>
+#include <stddef.h>
 #include <apic.h>
 #include <klib.h>
 #include <cpuid.h>
@@ -73,14 +75,14 @@ void lapic_send_ipi(uint8_t vector, uint8_t target_id) {
 
 /* Read from the `io_apic_num`'th I/O APIC as described by the MADT */
 uint32_t io_apic_read(size_t io_apic_num, uint32_t reg) {
-    volatile uint32_t *base = (volatile uint32_t *)madt_io_apics[io_apic_num]->addr;
+    volatile uint32_t *base = (volatile uint32_t *)(size_t)madt_io_apics[io_apic_num]->addr;
     *base = reg;
     return *(base + 4);
 }
 
 /* Write to the `io_apic_num`'th I/O APIC as described by the MADT */
 void io_apic_write(size_t io_apic_num, uint32_t reg, uint32_t data) {
-    volatile uint32_t *base = (volatile uint32_t *)madt_io_apics[io_apic_num]->addr;
+    volatile uint32_t *base = (volatile uint32_t *)(size_t)madt_io_apics[io_apic_num]->addr;
     *base = reg;
     *(base + 4) = data;
     return;
